Level-editor: Texture instead of cursor
Instead of a cursor there is the texture of the current terrain type, but I'm not content. The texture is an Image in an ImageView which is following the mouse. When resizing the brush, I can make the ImageView bigger, but not tell it to repeat the texture so that the texture is just enlarged and becomes very blurry.